set bug fix

+++ node0.hpp
@@ -53,6 +53,8 @@ class rbtree{
 			this->root = this->TNULL;
 		}
 
+		rbtree(const_reference rhs) : root(0), cmp(rhs.cmp) { *this = rhs; }
+
 		~rbtree() { delete_all_node(this->root); }
 
 		rbtree &operator=(rbtree const &rhs)
@@ -62,8 +64,10 @@ class rbtree{
 			this->con_alloc = rhs.con_alloc;
 			this->node_alloc = rhs.node_alloc;
 			this->cmp = rhs.cmp;
-
-			delete_all_node(this->root);
+			if (this -> root == NULL)
+				init_tree();
+			else 
+				delete_all_node(this->root);
 			if (rhs._size == 0)
 				this->root = this->TNULL;
 			else
@@ -422,7 +426,7 @@ class rbtree{
 	}
 	void init_tree() {
 		TNULL = node_alloc.allocate(1);
-		node_alloc.construct( TNULL, value_type());
+		node_alloc.construct( TNULL, Node<Value>());
 		TNULL -> color = 0;
 		TNULL -> is_null = 1;
 	}
